+++ client/internal/oicq/ecdh.go
@@ -1,5 +1,7 @@
 package oicq
 
+// from https://github.com/Mrs4s/MiraiGo/blob/master/client/internal/oicq/ecdh.go
+
 import (
 	"crypto/ecdh"
 	"crypto/md5"
